refactor: feature-gate parquet and vortex support at build time
Extracts format-agnostic conversion pieces (DNA packing, per-blob
decode, length-mode detection) out of the parquet module into a new
`crate::convert` module, then makes parquet and vortex optional
Cargo features.

Cargo features on `sracha-core`:

- `parquet` (default): pulls `arrow` + `parquet`, enables `crate::parquet`.
- `vortex` (default): pulls `vortex` + `vortex-btrblocks` + `futures`,
  enables `crate::vortex`.

Matching features propagate through `sracha`:

    cargo build                                      # both (default)
    cargo build --no-default-features --features parquet
    cargo build --no-default-features --features vortex
    cargo build --no-default-features                # neither; Convert
                                                     # subcommand is gone

All four combinations compile cleanly. `cargo test -p sracha-core`
passes (332 tests). `cargo clippy --workspace --all-targets
--all-features -- -D warnings` clean.

Biggest non-Cargo change: `crate::convert::decode` now owns
`DecodedBlob`, `SpotIter`/`ReadIter`, `decode_one_blob`,
`resolve_length_mode`, and `pack_sequence`; `crate::convert::schema`
owns the `DnaPacking` / `LengthMode` / `LengthModeChoice` enums.
Parquet module keeps the Arrow-schema builder and the parquet-specific
writer; vortex module keeps the native builder and the vortex-specific
writer — both now depend on `convert::*` instead of each other, so
enabling vortex no longer forces the parquet crate into the build.

CLI: `Convert` subcommand and its arg types are all `#[cfg]`-gated,
`--format` dynamically defaults to whichever format is compiled in,
and the parquet-only knobs (`--compression`, `--zstd-level`,
`--row-group-mib`) disappear when parquet is not enabled.
