+++ crates/sracha-core/src/parquet/mod.rs
@@ -5,10 +5,13 @@
 //! Apache Arrow / Parquet can match VDB's storage density when given access to
 //! the same domain tricks (2na DNA packing, fixed-width detection,
 //! dictionary-encoded quality).
+//!
+//! Format-agnostic pieces (DNA packing, per-blob decode, length-mode
+//! detection) live in [`crate::convert`]; this module only owns the
+//! Parquet-specific Arrow schema and writer.
 
-pub mod encoding;
 pub mod schema;
 pub mod writer;
 
-pub use schema::{DnaPacking, LengthMode};
+pub use crate::convert::schema::{DnaPacking, LengthMode};
 pub use writer::{ConvertConfig, ConvertStats, ParquetCompression, convert_sra_to_parquet};
